requirements.txt
python-binance
typer
python-dotenv
pydantic